requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
comm
cryptography==45.0.5
debugpy
decorator
distro==1.9.0
exceptiongroup
executing
fastapi==0.116.1
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jedi
jupyter_client
jupyter_core
matplotlib-inline
nest_asyncio
oauthlib==3.3.1
packaging
parso
pexpect
pickleshare
platformdirs
prompt_toolkit
proto-plus==1.26.1
protobuf==6.31.1
psutil
ptyprocess
pure_eval
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments
pyparsing==3.2.3
python-dateutil
python-dotenv==1.1.1
pyzmq
requests==2.32.4
requests-oauthlib==2.0.0
retell-sdk==4.40.0
rsa==4.9.1
setuptools==78.1.1
six
sniffio==1.3.1
stack_data
starlette==0.47.2
tornado
traitlets
typing-inspection==0.4.1
typing_extensions
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth
wheel==0.45.1
zipp
